React: link to tables from info dialogs (#181)
When the user opens a detail dialog about any given element in a table (datasets, analyses, participants), and they open one of the listed items from a different table (eg. datasets from a participant, or analyses from a dataset), clicking the 'Open in table' button will take them to that other table, and change the ID filter so that the item they clicked is the only one present.

This functionality is mostly handled by react-router. Any links to `/<tablename>/:id` will direct them to the specified table, with the specified ID in the table's ID filter.

Closes #163