React: fix groups mutations to only update group list (#375)
Invalidating queries with option `{ exact: true }` ensures that only queries that exactly match the given key will be invalidated, instead of all queries that contain it. The group create, update, delete hooks now use this option so that only the /api/groups query is invalidated, and not other /api/groups/:id queries that aren't affected by the action.