React: fix material-table filter reset bug caused by rerenders (#473)
It seems that the MaterialTable instance in `DatasetTable.tsx` was rerendering when `api/unlinked` was fetched in the background, causing filters on the table to reset. The table was rerendering b/c some of its props were inline functions which, according to react's equivalency checks, are technically 'new' on every render. Defining these functions outside of the component seems to have fixed the issue on my tests.