React: support larger page sizes and retrieving all rows (#524)
This was quite the adventure. At least it wasn't hair-ripping or frustrating.

Changes the default page size to 20 and the available page sizes to 20, 50, 100, and all. 
Implements support for fetching all rows at once, since it was benchmarked to take at most five seconds on my internet (for now) with the datasets table. Requires a MaterialTablePaginationOverride and an `any` cast to work correctly.
If there are fewer rows available than the current page size, the space will no longer be filled with blanks. This happens to be needed for the above to work correctly.

Refactor: add MaterialTablePrimary for common props between the three tables
Refactor: remove unused Analyses/components/DatasetTable

The rest should be documented in comments. (Some based on research in mbrn/material-table#1685 and the references there.)