Flask: don't update timestamp if there are no changes (#929)
When we save changes in the modals and if there are no changes in the field values, the `updated` timestamp would still change. This is caused by either a nonnull `library_prep_date` field or a nonnull `month_of_birth` field. When the backend receives the JSON request, the values for these fields are a string. However, the values of these fields are stored in the database as a date object. Thus, when we commit the session, SQLAlchemy will think that there's a change in value and update the timestamp. 

The solution here is to convert the values from `str` type to `datetime` type in the backend.