Flask: upgrade to python 3.8 (again) (#1042)
This PR is a continuation of #1028, reverted #1031. In production, importing our main wsgi module raised an error because `schemas.py` tries to `import typing_extensions`. This is not a standard library, but a transitive dependency of black, only available in dev. Furthermore, that import wasn't actually used in the file.

With this PR, simulating the production environment with `docker-compose up app_gunicorn` raised no errors.